feat: add CustomerDashboardForAdmin to support importing server components
